refactor: restructure application to follow MVC pattern
